Add files via upload
The STL files of the mold for the 0-Is (LST-G), 0-IIa+Is (LST-G), 0-IIa+IIc (LST-NG) polyp simulator (all three morphologies are combined into one mold), the 0-IIa (LST-NG) polyp simulator, the 0-Is (LST-NG) polyp simulator, and the 1sp polyp simulator are attached here. Print in 0.3mm or higher resolution in PLA plastic.